agent-orchestration-service/core/pyproject.toml
[project]
name = "common-ground-agent-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.8.0",
    "beautifulsoup4>=4.9.3",
    "openai>=1.0.0",
    "litellm>=1.30.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.0",
    "python-json-logger>=2.0.7",
    "requests>=2.28.0",
    "markdown>=3.4.0",
    "coolname>=1.1.0",
    "numpy>=2",  # 恢复numpy>=2，接受新版本
    "numba>=0.60.0",  # 允许更新版本以兼容numpy 2.x
    "json-repair==0.40.0",
    "pyyaml>=6.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "websockets>=12.0",
    "pygls==1.3.1",
    "lsprotocol==2023.0.1",
    "google-genai",
    "lark>=1.2.2",
    "querystring-parser>=1.2.4",
    "aiosqlite>=0.21.0",
    "aiofiles>=24.1.0",
    "mcp>=0.0.1",
    "duckdb==1.3.1",
    "colorama>=0.4.6",
    "watchdog",
    "cleantext>=1.1.0",
    "duckduckgo-search>=6.0.0",
    "beautifulsoup4>=4.9.3",
    "langchain-community>=0.0.13",
    "langchain-text-splitters>=0.0.1",
    "langchain-openai>=0.0.2",
    "pymupdf>=1.23.0",
    "tavily-python>=0.3.0",
    "yt-dlp>=2023.12.30",
    "markitdown[all]==0.1.2",
    "pocketflow>=0.0.1",
    "fastmcp>=0.0.1",
    "light-embed @ git+https://github.com/chux0519/light-embed@b98fc0f1262e0fa4a140f9b2a437b83f79d4158e",
    # 使用预编译版本的faiss避免编译问题
    "faiss-cpu>=1.8.0; platform_machine != 'arm64'",
    "faiss-cpu>=1.7.4; platform_machine == 'arm64'",
]
